This is a Simple Python Program which is used to analyze a particular chat under various sub-heads.
This is a Simple Python Program which is used to analyze a particular
chat under various sub-heads.
The given is an interpreter based application or Python shell program
which tells us the most active member, most active time, number of
messages, etc about a particular chat of a group or a person.